Metal advection (#8)
* first pass at advecting metals

* default cooling should not be not initialized when USE_GRACKLE.This is a major todo...

* Proper advection of metals.

The Metallicity is stored as teh 0th element in the PScalars. At the moment total metal abundance is tracked.
There is a new input paramter `InitMetallicityinSolar` which takes in the inital metallicity for the gas cells in units of SOLAR_ABUNDANCE (set to be 0.0134)

* comments; plus cleanup

* rename Metallicity to Metals in the SphP struct

So that the compiler doesn't confuse with the Metallicity in the CELib
structs....